Move to navigate_to_pose/
