+++ nav2_bt_navigator/CMakeLists.txt
@@ -22,17 +22,17 @@ include_directories(
   include
 )
 
-set(executable_name bt_navigator)
+set(navigate_to_pose_executable_name bt_navigator)
 
-add_executable(${executable_name}
-  src/main.cpp
+add_executable(${navigate_to_pose_executable_name}
+  src/navigate_to_pose/main.cpp
 )
 
-set(library_name ${executable_name}_core)
+set(library_name ${navigate_to_pose_executable_name}_core)
 
 add_library(${library_name} SHARED
   src/bt_navigator.cpp
-  src/navigate_to_pose_bt_navigator.cpp
+  src/navigate_to_pose/bt_navigator.cpp
   src/ros_topic_logger.cpp
 )
 
@@ -51,11 +51,11 @@ set(dependencies
   tf2_ros
 )
 
-ament_target_dependencies(${executable_name}
+ament_target_dependencies(${navigate_to_pose_executable_name}
   ${dependencies}
 )
 
-target_link_libraries(${executable_name} ${library_name})
+target_link_libraries(${navigate_to_pose_executable_name} ${library_name})
 
 ament_target_dependencies(${library_name}
   ${dependencies}
@@ -67,7 +67,7 @@ install(TARGETS ${library_name}
   RUNTIME DESTINATION bin
 )
 
-install(TARGETS ${executable_name}
+install(TARGETS ${navigate_to_pose_executable_name}
   RUNTIME DESTINATION lib/${PROJECT_NAME}
 )
